Publicize buffer algorithms previously in ntsu_buffer to ntsa_buffer
