+++ groups/nts/ntsu/package/ntsu.mem
@@ -1,5 +1,4 @@
 ntsu_adapterutil
-ntsu_bufferutil
 ntsu_zerocopyutil
 ntsu_resolverutil
 ntsu_socketutil
