+++ targets.cmake
@@ -216,13 +216,11 @@ if (${NTF_BUILD_WITH_NTS})
     )
 
     ntf_component(NAME ntsu_adapterutil)
-    ntf_component(NAME ntsu_bufferutil)
     ntf_component(NAME ntsu_zerocopyutil)
     ntf_component(NAME ntsu_resolverutil)
     ntf_component(NAME ntsu_socketutil)
     ntf_component(NAME ntsu_socketoptionutil)
     ntf_component(NAME ntsu_timestamputil)
-
 
     ntf_package_end(NAME ntsu)
